Split PET renderer into focused modules
