+++ src/web/src/components/home-pet/cloud-code-monster-pet-constants.ts
@@ -1,0 +1,26 @@
+export const CLOUD_CODE_MONSTER_ACTIVITY_REFRESH_MS = 3 * 60 * 1000;
+export const CLOUD_CODE_MONSTER_STORAGE_KEY = "alook-cloud-code-monster-pet-activity-v1";
+export const CLOUD_CODE_MONSTER_PRESET_STORAGE_KEY =
+  "alook-cloud-code-monster-pet-preset-v1";
+export const CLOUD_CODE_MONSTER_PRESET_CHANGED_EVENT =
+  "alook-cloud-code-monster-pet-preset-changed";
+export const CLOUD_CODE_MONSTER_REACTION_MS = 900;
+export const CLOUD_CODE_MONSTER_SHAKE_REACTION_MS = 680;
+export const CLOUD_CODE_MONSTER_FAINT_MS = 10_000;
+export const CLOUD_CODE_MONSTER_VIOLENT_DRAG_MIN_DISTANCE = 24;
+export const CLOUD_CODE_MONSTER_VIOLENT_DRAG_STRONG_DISTANCE = 56;
+export const CLOUD_CODE_MONSTER_VIOLENT_DRAG_MAX_ELAPSED_MS = 70;
+export const CLOUD_CODE_MONSTER_VIOLENT_DRAG_MIN_SPEED = 1.2;
+export const CLOUD_CODE_MONSTER_VIOLENT_REVERSAL_MIN_SPEED = 1.5;
+export const CLOUD_CODE_MONSTER_FAINT_DRAG_MIN_DISTANCE = 72;
+export const CLOUD_CODE_MONSTER_FAINT_DRAG_MAX_ELAPSED_MS = 55;
+export const CLOUD_CODE_MONSTER_FAINT_DRAG_MIN_SPEED = 2.2;
+export const CLOUD_CODE_MONSTER_FAINT_REVERSAL_MIN_SPEED = 2.75;
+export const CLOUD_CODE_MONSTER_FAINT_EVENT_WINDOW_MS = 1_300;
+export const CLOUD_CODE_MONSTER_FAINT_MIN_EVENTS = 6;
+export const CLOUD_CODE_MONSTER_FAINT_MIN_SPAN_MS = 450;
+export const CLOUD_CODE_MONSTER_AUTO_WALK_STEP_MS = 80;
+export const CLOUD_CODE_MONSTER_AUTO_WALK_SPEED = 2.7;
+export const CLOUD_CODE_MONSTER_PEEK_MS = 4_200;
+export const CLOUD_CODE_MONSTER_PEEK_INTERVAL_MS = 11_000;
+export const CLOUD_CODE_MONSTER_SIZE = { width: 82, height: 82 };
