+++ src/web/src/components/home-pet/cloud-code-monster-pet-presets.ts
@@ -1,0 +1,592 @@
+import {
+  CLOUD_CODE_MONSTER_PRESET_CHANGED_EVENT,
+  CLOUD_CODE_MONSTER_PRESET_STORAGE_KEY,
+} from "./cloud-code-monster-pet-constants";
+import type { CloudCodeMonsterPetPreset } from "./cloud-code-monster-pet-types";
+
+const CLOUD_CODE_MONSTER_PET_PRESET_BASES: CloudCodeMonsterPetPreset[] = [
+  {
+    id: "pet-01",
+    name: "Claude Pixel",
+    group: "Codex",
+    feature: "square",
+    bodyTop: "#e0845f",
+    body: "#d87352",
+    bodyDark: "#cf684a",
+    bodyLight: "#e88d68",
+    bodySideLight: "#df7d5a",
+    bodySideDark: "#c86347",
+    accent: "#d7654a",
+    accessory: "#f1bc62",
+    eye: "#12100f",
+    highlight: "#f6e7d7",
+  },
+  {
+    id: "pet-02",
+    name: "Blue Pocket Bot",
+    group: "Gadget Buddy",
+    feature: "bell",
+    bodyTop: "#4aa8d8",
+    body: "#2f8fc4",
+    bodyDark: "#2674a4",
+    bodyLight: "#68bce5",
+    bodySideLight: "#48a8d5",
+    bodySideDark: "#1f6591",
+    accent: "#d84e4a",
+    accessory: "#f4c94f",
+    eye: "#10141a",
+    highlight: "#f7fbff",
+    facePatch: "#f6fbff",
+    cheek: "#d84e4a",
+  },
+  {
+    id: "pet-03",
+    name: "Spark Arcade Pal",
+    group: "Electric Mascot",
+    feature: "bolt",
+    bodyTop: "#f2d45a",
+    body: "#e6bd37",
+    bodyDark: "#c89a2a",
+    bodyLight: "#ffe37a",
+    bodySideLight: "#f0cc4c",
+    bodySideDark: "#b58724",
+    accent: "#2f2520",
+    accessory: "#df5145",
+    eye: "#17120e",
+    highlight: "#fff7c7",
+    cheek: "#e94d43",
+  },
+  {
+    id: "pet-04",
+    name: "Pink Star Puff",
+    group: "Star Puff",
+    feature: "star",
+    bodyTop: "#eba0b7",
+    body: "#df7fa0",
+    bodyDark: "#c76589",
+    bodyLight: "#f2b6c8",
+    bodySideLight: "#e993ac",
+    bodySideDark: "#b95778",
+    accent: "#c84d64",
+    accessory: "#f0d862",
+    eye: "#171016",
+    highlight: "#fff0f6",
+    cheek: "#e75c7a",
+  },
+  {
+    id: "pet-05",
+    name: "Leaf Sprite",
+    group: "Forest Starter",
+    feature: "leaf",
+    bodyTop: "#76aa7b",
+    body: "#5f9666",
+    bodyDark: "#4d7c54",
+    bodyLight: "#91bd8f",
+    bodySideLight: "#72a875",
+    bodySideDark: "#426d49",
+    accent: "#2f6f46",
+    accessory: "#d6e78a",
+    eye: "#11160f",
+    highlight: "#f0f8e4",
+    facePatch: "#dceec2",
+  },
+  {
+    id: "pet-06",
+    name: "Ember Starter",
+    group: "Fire Buddy",
+    feature: "flame",
+    bodyTop: "#df8755",
+    body: "#ce6b3f",
+    bodyDark: "#af5333",
+    bodyLight: "#ee9d6b",
+    bodySideLight: "#dd7b4c",
+    bodySideDark: "#9a482c",
+    accent: "#8d3d2f",
+    accessory: "#f3c34b",
+    eye: "#17100c",
+    highlight: "#ffe8d4",
+    cheek: "#e45842",
+  },
+  {
+    id: "pet-07",
+    name: "Ripple Buddy",
+    group: "Water Buddy",
+    feature: "fins",
+    bodyTop: "#82b5d6",
+    body: "#629dcc",
+    bodyDark: "#4f82ad",
+    bodyLight: "#9bc8e5",
+    bodySideLight: "#73add6",
+    bodySideDark: "#44749b",
+    accent: "#5f6f8d",
+    accessory: "#ead59a",
+    eye: "#10151b",
+    highlight: "#eff8ff",
+    facePatch: "#d9eef8",
+  },
+  {
+    id: "pet-08",
+    name: "Moon Alley Cat",
+    group: "Moon Cat",
+    feature: "moon",
+    bodyTop: "#4f4b69",
+    body: "#3f3a59",
+    bodyDark: "#332f49",
+    bodyLight: "#67607e",
+    bodySideLight: "#554f6d",
+    bodySideDark: "#2b283e",
+    accent: "#e5c85c",
+    accessory: "#d66f91",
+    eye: "#f5e8a7",
+    highlight: "#fff6c8",
+    cheek: "#cf6d8f",
+  },
+  {
+    id: "pet-09",
+    name: "Spore Runner",
+    group: "Mushroom Runner",
+    feature: "mushroom",
+    bodyTop: "#e0524f",
+    body: "#f0dcc6",
+    bodyDark: "#d7c0aa",
+    bodyLight: "#f7ead9",
+    bodySideLight: "#f1dfcb",
+    bodySideDark: "#c7aa95",
+    accent: "#b83736",
+    accessory: "#ffffff",
+    eye: "#17120f",
+    highlight: "#fff8ea",
+  },
+  {
+    id: "pet-10",
+    name: "Dash Quill",
+    group: "Speed Mascot",
+    feature: "spin",
+    bodyTop: "#4e83d0",
+    body: "#356dbd",
+    bodyDark: "#2a579b",
+    bodyLight: "#6896de",
+    bodySideLight: "#477dd0",
+    bodySideDark: "#244c88",
+    accent: "#f2c45b",
+    accessory: "#f2ece1",
+    eye: "#10131a",
+    highlight: "#eef4ff",
+    facePatch: "#ead3b9",
+  },
+  {
+    id: "pet-11",
+    name: "Lemon Chomper",
+    group: "Arcade Chomper",
+    feature: "chomp",
+    bodyTop: "#f0ce45",
+    body: "#e5ba2f",
+    bodyDark: "#c89924",
+    bodyLight: "#ffe26c",
+    bodySideLight: "#efc73f",
+    bodySideDark: "#b18520",
+    accent: "#1f1b18",
+    accessory: "#f2845f",
+    eye: "#15110d",
+    highlight: "#fff2b6",
+  },
+  {
+    id: "pet-12",
+    name: "Lantern Ghost",
+    group: "Round Ghost",
+    feature: "ghost",
+    bodyTop: "#d9dff0",
+    body: "#c4cee6",
+    bodyDark: "#aab6d2",
+    bodyLight: "#edf2ff",
+    bodySideLight: "#d3dced",
+    bodySideDark: "#9aa8c6",
+    accent: "#6d7da8",
+    accessory: "#e77c9c",
+    eye: "#151926",
+    highlight: "#ffffff",
+  },
+  {
+    id: "pet-13",
+    name: "Red Cap Jumper",
+    group: "Platform Pal",
+    feature: "cap",
+    bodyTop: "#d84f45",
+    body: "#c5a17d",
+    bodyDark: "#a98263",
+    bodyLight: "#e0ba94",
+    bodySideLight: "#d3ad86",
+    bodySideDark: "#936f54",
+    accent: "#2c4f99",
+    accessory: "#f6e9d1",
+    eye: "#16110d",
+    highlight: "#fff2df",
+  },
+  {
+    id: "pet-14",
+    name: "Honey Cub",
+    group: "Classic Toy",
+    feature: "ears",
+    bodyTop: "#e4b25a",
+    body: "#d0993e",
+    bodyDark: "#ad7d30",
+    bodyLight: "#efc46d",
+    bodySideLight: "#dba84d",
+    bodySideDark: "#996d2a",
+    accent: "#b54d3f",
+    accessory: "#f4d991",
+    eye: "#17110d",
+    highlight: "#fff0c8",
+    facePatch: "#f2d083",
+  },
+  {
+    id: "pet-15",
+    name: "Ribbon Puff",
+    group: "Bow Buddy",
+    feature: "bow",
+    bodyTop: "#f2f0e8",
+    body: "#e3ded4",
+    bodyDark: "#c9c2b7",
+    bodyLight: "#ffffff",
+    bodySideLight: "#eee9df",
+    bodySideDark: "#bab2a7",
+    accent: "#d84b5b",
+    accessory: "#f3cc56",
+    eye: "#15110f",
+    highlight: "#ffffff",
+    cheek: "#df6d80",
+  },
+  {
+    id: "pet-16",
+    name: "Cozy Hood Cub",
+    group: "Hooded Mascot",
+    feature: "hood",
+    bodyTop: "#ef9eba",
+    body: "#e681a4",
+    bodyDark: "#c96789",
+    bodyLight: "#f5b6ca",
+    bodySideLight: "#ee94b2",
+    bodySideDark: "#b85878",
+    accent: "#f0d6de",
+    accessory: "#f3d66b",
+    eye: "#171017",
+    highlight: "#fff0f6",
+    facePatch: "#fff6f2",
+  },
+  {
+    id: "pet-17",
+    name: "Masked Beetle",
+    group: "Bug Hero",
+    feature: "mask",
+    bodyTop: "#6e617d",
+    body: "#554761",
+    bodyDark: "#41374b",
+    bodyLight: "#837491",
+    bodySideLight: "#655570",
+    bodySideDark: "#382f41",
+    accent: "#cf6f8e",
+    accessory: "#f0d46a",
+    eye: "#f7edf4",
+    highlight: "#fff7ff",
+    facePatch: "#ece0ea",
+  },
+  {
+    id: "pet-18",
+    name: "Round Forest Pal",
+    group: "Forest Neighbor",
+    feature: "ears",
+    bodyTop: "#8c8d7d",
+    body: "#74766a",
+    bodyDark: "#606257",
+    bodyLight: "#a0a293",
+    bodySideLight: "#838578",
+    bodySideDark: "#55574d",
+    accent: "#d9d3b8",
+    accessory: "#c6d68c",
+    eye: "#121411",
+    highlight: "#efead8",
+    facePatch: "#dad4ba",
+  },
+  {
+    id: "pet-19",
+    name: "Dust Puff",
+    group: "Tiny Soot",
+    feature: "soot",
+    bodyTop: "#34343a",
+    body: "#25262c",
+    bodyDark: "#1d1e24",
+    bodyLight: "#4a4b52",
+    bodySideLight: "#383941",
+    bodySideDark: "#171820",
+    accent: "#6c6f7a",
+    accessory: "#f0f0d8",
+    eye: "#f7f7e8",
+    highlight: "#ffffff",
+  },
+  {
+    id: "pet-20",
+    name: "Straw Field Pal",
+    group: "Harvest Buddy",
+    feature: "straw",
+    bodyTop: "#d96a50",
+    body: "#c65342",
+    bodyDark: "#a74336",
+    bodyLight: "#e47d64",
+    bodySideLight: "#d0614d",
+    bodySideDark: "#933a30",
+    accent: "#2f5f93",
+    accessory: "#e8c866",
+    eye: "#15100d",
+    highlight: "#ffe6d9",
+  },
+  {
+    id: "pet-21",
+    name: "Shadow Ninja Bean",
+    group: "Ninja Mascot",
+    feature: "ninja",
+    bodyTop: "#e7a04d",
+    body: "#d88432",
+    bodyDark: "#b96c27",
+    bodyLight: "#efb464",
+    bodySideLight: "#df943e",
+    bodySideDark: "#a25d22",
+    accent: "#2d415c",
+    accessory: "#ded5c6",
+    eye: "#15100c",
+    highlight: "#ffe7c8",
+    facePatch: "#f3c78d",
+  },
+  {
+    id: "pet-22",
+    name: "Pearl Quest Cub",
+    group: "Pearl Hero",
+    feature: "pearl",
+    bodyTop: "#e3a54e",
+    body: "#ce8438",
+    bodyDark: "#ac682e",
+    bodyLight: "#efbc68",
+    bodySideLight: "#db9345",
+    bodySideDark: "#985a28",
+    accent: "#476ca6",
+    accessory: "#f0d464",
+    eye: "#15100c",
+    highlight: "#ffe8c4",
+  },
+  {
+    id: "pet-23",
+    name: "Wand Starling",
+    group: "Magic Mascot",
+    feature: "wand",
+    bodyTop: "#f0a6c4",
+    body: "#df87ad",
+    bodyDark: "#c26b92",
+    bodyLight: "#f5bad2",
+    bodySideLight: "#e99ab9",
+    bodySideDark: "#ad5d82",
+    accent: "#d8b34a",
+    accessory: "#e9d7ff",
+    eye: "#171017",
+    highlight: "#fff0f7",
+    cheek: "#e76791",
+  },
+  {
+    id: "pet-24",
+    name: "Mecha Bean",
+    group: "Tiny Mecha",
+    feature: "mecha",
+    bodyTop: "#d7dbe2",
+    body: "#bdc3ce",
+    bodyDark: "#9ea6b3",
+    bodyLight: "#edf0f5",
+    bodySideLight: "#cbd1da",
+    bodySideDark: "#8d96a5",
+    accent: "#d44d4a",
+    accessory: "#f0c748",
+    eye: "#151923",
+    highlight: "#ffffff",
+  },
+  {
+    id: "pet-25",
+    name: "Bubble Slimelet",
+    group: "Slime Buddy",
+    feature: "slime",
+    bodyTop: "#79b8e8",
+    body: "#5aa0d8",
+    bodyDark: "#4384ba",
+    bodyLight: "#96cdf0",
+    bodySideLight: "#70b3e3",
+    bodySideDark: "#3975a6",
+    accent: "#2a658f",
+    accessory: "#f3f7ff",
+    eye: "#101722",
+    highlight: "#f5fbff",
+  },
+  {
+    id: "pet-26",
+    name: "Ink Courier",
+    group: "Ink Runner",
+    feature: "ink",
+    bodyTop: "#75c8bb",
+    body: "#54afa2",
+    bodyDark: "#3f9188",
+    bodyLight: "#90d9ce",
+    bodySideLight: "#67c1b5",
+    bodySideDark: "#377f77",
+    accent: "#d7668f",
+    accessory: "#e4f2ef",
+    eye: "#101716",
+    highlight: "#effffb",
+  },
+  {
+    id: "pet-27",
+    name: "Drum Marcher",
+    group: "Rhythm Pal",
+    feature: "drum",
+    bodyTop: "#ef7f5f",
+    body: "#df6349",
+    bodyDark: "#bd4d3b",
+    bodyLight: "#f39a7b",
+    bodySideLight: "#e77358",
+    bodySideDark: "#a84434",
+    accent: "#4f78b6",
+    accessory: "#f3e4c6",
+    eye: "#15100d",
+    highlight: "#fff0e4",
+  },
+  {
+    id: "pet-28",
+    name: "Sprout Forestling",
+    group: "Sprout Buddy",
+    feature: "sprout",
+    bodyTop: "#a58f5d",
+    body: "#8f7548",
+    bodyDark: "#755f3a",
+    bodyLight: "#b9a26e",
+    bodySideLight: "#9b8354",
+    bodySideDark: "#655233",
+    accent: "#5d8b56",
+    accessory: "#dfc86e",
+    eye: "#15110c",
+    highlight: "#f8ecc0",
+    facePatch: "#d8be71",
+  },
+  {
+    id: "pet-29",
+    name: "Frost Star Pup",
+    group: "Ice Pal",
+    feature: "star",
+    bodyTop: "#8aa7d8",
+    body: "#6e91c8",
+    bodyDark: "#5877aa",
+    bodyLight: "#a1bce5",
+    bodySideLight: "#7f9fd3",
+    bodySideDark: "#4e6898",
+    accent: "#f0cf5b",
+    accessory: "#f3edf5",
+    eye: "#10151c",
+    highlight: "#eff5ff",
+    cheek: "#d66d86",
+  },
+  {
+    id: "pet-30",
+    name: "Tiny Quest Cub",
+    group: "Quest Mascot",
+    feature: "bow",
+    bodyTop: "#efa2c8",
+    body: "#df84b2",
+    bodyDark: "#c46999",
+    bodyLight: "#f5b9d5",
+    bodySideLight: "#eb96c2",
+    bodySideDark: "#ad5a86",
+    accent: "#cf4d78",
+    accessory: "#f0d66a",
+    eye: "#171016",
+    highlight: "#fff0f8",
+    cheek: "#df688f",
+  },
+];
+
+const INSPIRED_PRESET_OVERRIDES = [
+  { id: "pet-02", name: "Pocket Bell Bot", group: "Inspired Mascots", feature: "bell", shape: "doraemon" },
+  { id: "pet-03", name: "Spark Bolt Pal", group: "Inspired Mascots", feature: "bolt", shape: "pikachu" },
+  { id: "pet-04", name: "Star Puff", group: "Inspired Mascots", feature: "star", shape: "kirby" },
+  { id: "pet-05", name: "Leaf Bud", group: "Inspired Mascots", feature: "leaf", shape: "bulbasaur" },
+  { id: "pet-06", name: "Tiny Flame", group: "Inspired Mascots", feature: "flame", shape: "charmander" },
+  { id: "pet-07", name: "Shell Fin", group: "Inspired Mascots", feature: "fins", shape: "squirtle" },
+  { id: "pet-08", name: "Block Builder", group: "Inspired Mascots", feature: "visor", shape: "minecraft-steve" },
+  { id: "pet-09", name: "Block Hiss", group: "Inspired Mascots", feature: "mask", shape: "minecraft-creeper" },
+  { id: "pet-10", name: "Block Walker", group: "Inspired Mascots", feature: "visor", shape: "minecraft-zombie" },
+  { id: "pet-11", name: "Mushroom Pal", group: "Inspired Mascots", feature: "mushroom", shape: "toad" },
+  { id: "pet-12", name: "Blue Spin", group: "Inspired Mascots", feature: "spin", shape: "sonic" },
+  { id: "pet-13", name: "Arcade Chomp", group: "Inspired Mascots", feature: "chomp", shape: "pacman" },
+  { id: "pet-14", name: "Peek Ghost", group: "Inspired Mascots", feature: "ghost", shape: "boo" },
+  { id: "pet-15", name: "Red Cap Jumper", group: "Inspired Mascots", feature: "cap", shape: "mario" },
+  { id: "pet-16", name: "Honey Cub", group: "Inspired Mascots", feature: "ears", shape: "pooh" },
+  { id: "pet-17", name: "Ribbon Cat", group: "Inspired Mascots", feature: "bow", shape: "hello-kitty" },
+  { id: "pet-18", name: "Cozy Hood Bunny", group: "Inspired Mascots", feature: "hood", shape: "my-melody" },
+  { id: "pet-19", name: "Imp Hood", group: "Inspired Mascots", feature: "mask", shape: "kuromi" },
+  { id: "pet-20", name: "Forest Neighbor", group: "Inspired Mascots", feature: "ears", shape: "totoro" },
+  { id: "pet-21", name: "Dust Puff", group: "Inspired Mascots", feature: "soot", shape: "soot-sprite" },
+  { id: "pet-22", name: "Straw Voyager", group: "Inspired Mascots", feature: "straw", shape: "luffy" },
+  { id: "pet-23", name: "Leaf Runner", group: "Inspired Mascots", feature: "ninja", shape: "naruto" },
+  { id: "pet-24", name: "Energy Pearl", group: "Inspired Mascots", feature: "pearl", shape: "goku" },
+  { id: "pet-25", name: "Moon Wand", group: "Inspired Mascots", feature: "wand", shape: "sailor-moon" },
+  { id: "pet-26", name: "Mecha Guard", group: "Inspired Mascots", feature: "mecha", shape: "gundam" },
+  { id: "pet-27", name: "Teardrop Slime", group: "Inspired Mascots", feature: "slime", shape: "dragon-quest-slime" },
+  { id: "pet-28", name: "Ink Runner", group: "Inspired Mascots", feature: "ink", shape: "inkling" },
+  { id: "pet-29", name: "Little Beagle", group: "Inspired Mascots", feature: "ears", shape: "snoopy" },
+  { id: "pet-30", name: "Tiny Antler", group: "Inspired Mascots", feature: "horns", shape: "chopper" },
+] satisfies Array<
+  Pick<CloudCodeMonsterPetPreset, "id" | "name" | "group" | "feature" | "shape">
+>;
+
+const INSPIRED_PRESET_OVERRIDE_BY_ID = new Map(
+  INSPIRED_PRESET_OVERRIDES.map((preset) => [preset.id, preset])
+);
+
+export const CLOUD_CODE_MONSTER_PET_PRESETS: CloudCodeMonsterPetPreset[] =
+  CLOUD_CODE_MONSTER_PET_PRESET_BASES.map((preset) => ({
+    ...preset,
+    ...INSPIRED_PRESET_OVERRIDE_BY_ID.get(preset.id),
+  }));
+
+export function getCloudCodeMonsterPreset(presetId?: string | null) {
+  return (
+    CLOUD_CODE_MONSTER_PET_PRESETS.find((preset) => preset.id === presetId) ??
+    CLOUD_CODE_MONSTER_PET_PRESETS[0]!
+  );
+}
+
+export function readCloudCodeMonsterPetPresetId() {
+  if (typeof localStorage === "undefined") {
+    return CLOUD_CODE_MONSTER_PET_PRESETS[0]!.id;
+  }
+
+  try {
+    return getCloudCodeMonsterPreset(
+      localStorage.getItem(CLOUD_CODE_MONSTER_PRESET_STORAGE_KEY)
+    ).id;
+  } catch {
+    return CLOUD_CODE_MONSTER_PET_PRESETS[0]!.id;
+  }
+}
+
+export function writeCloudCodeMonsterPetPresetId(presetId: string) {
+  const nextPreset = getCloudCodeMonsterPreset(presetId);
+
+  if (typeof localStorage !== "undefined") {
+    try {
+      localStorage.setItem(CLOUD_CODE_MONSTER_PRESET_STORAGE_KEY, nextPreset.id);
+    } catch {}
+  }
+
+  if (typeof window !== "undefined") {
+    window.dispatchEvent(
+      new CustomEvent(CLOUD_CODE_MONSTER_PRESET_CHANGED_EVENT, {
+        detail: { presetId: nextPreset.id },
+      })
+    );
+  }
+
+  return nextPreset.id;
+}
